+++ example/todo/web/const/TodoDataParams.dart
@@ -1,6 +1,8 @@
 class TodoDataParams {
   static const String
-    LIST = 'DATA_TODO_LIST'
+    LIST = 'DATA_TODO__ALL_LIST',
+    COUNT = 'DATA_TODO__NOT_COMPLETED_COUNT',
+    FILTER = 'DATA_TODO__FILTER'
   ;
 }