+++ example/todo/web/model/TodoModel.dart
@@ -1,79 +1,146 @@
 import 'dart:convert';
 import 'dart:html';
-
 import 'package:wire/wire.dart';
-
 import '../const/TodoDataParams.dart';
+import '../const/TodoFilterValues.dart';
 import 'ITodoModel.dart';
-import 'vos/TodoVO.dart';
+import 'vo/TodoVO.dart';
 
 class TodoModel implements ITodoModel {
   static const String LOCAL_STORAGE_KEY = 'todo-mvc-dart-wire';
 
   TodoModel() {
     var idsList = <String>[];
+    var notCompletedCount = 0;
     if (window.localStorage.containsKey(LOCAL_STORAGE_KEY)) {
       try {
         jsonDecode(window.localStorage[LOCAL_STORAGE_KEY]).forEach((obj){
-          var todoVO = TodoVO.fromJson(obj);
-          Wire.data(todoVO.id, todoVO);
-          idsList.add(todoVO.id);
+          if (obj != null) {
+            var todoVO = TodoVO.fromJson(obj);
+            Wire.data(todoVO.id, todoVO);
+            idsList.add(todoVO.id);
+            if (!todoVO.completed) notCompletedCount++;
+          }
         });
       } catch (e) {
         print('Error loading form local storage: ' + e.toString());
       }
     }
     Wire.data(TodoDataParams.LIST, idsList);
+    Wire.data(TodoDataParams.COUNT, notCompletedCount);
   }
 
   @override
   TodoVO create(String text) {
     final time = DateTime.now().millisecondsSinceEpoch;
     final id = time.toString();
-    final vo = TodoVO(id, text, false);
-    final data = Wire.data(TodoDataParams.LIST);
-    final list = data.value as List;
+    final todoVO = TodoVO(id, text, false);
+    final listData = Wire.data(TodoDataParams.LIST);
+    final todoList = listData.value as List;
+    final count = Wire.data(TodoDataParams.COUNT).value as int;
 
-    list.add(vo.id);
-    Wire.data(vo.id, vo);
-    Wire.data(TodoDataParams.LIST, list);
+    todoList.add(todoVO.id);
+    Wire.data(todoVO.id, todoVO);
+    Wire.data(TodoDataParams.LIST, todoList);
+    Wire.data(TodoDataParams.COUNT, count + 1);
 
     _save();
 
-    print('> TodoModel -> create: ' + vo.id + ' - ' + vo.text);
-    return vo;
+    print('> TodoModel -> created: ' + todoVO.id + ' - ' + todoVO.text);
+    return todoVO;
   }
+
   @override
   TodoVO remove(String id) {
-    final list = Wire.data(TodoDataParams.LIST).value as List;
-    final data = Wire.data(id);
-    final vo = data.value;
+    final todoList = Wire.data(TodoDataParams.LIST).value as List;
+    final count = Wire.data(TodoDataParams.COUNT).value as int;
+    final todoWireData = Wire.data(id);
+    final todoVO = todoWireData.value as TodoVO;
 
-    list.remove(id);
-    data.remove();
+    todoList.remove(id);
+    todoWireData.remove();
+
+    if (todoVO.completed == false) {
+      Wire.data(TodoDataParams.COUNT, count -1);
+    }
 
     _save();
 
-    print('> TodoModel -> remove: ' + id);
-    return vo;
+    print('> TodoModel -> removed: ' + id);
+    return todoVO;
   }
 
   @override
   TodoVO update(String id, String text) {
+    final todoWireData = Wire.data(id);
+    final todoVO = todoWireData.value as TodoVO;
+    todoVO.text = text;
+    Wire.data(id, todoVO);
     _save();
-    return null;
+
+    print('> TodoModel -> updated: ' + todoVO.id + ' - ' + todoVO.text);
+    return todoVO;
   }
+
   @override
   TodoVO toggle(String id) {
+    final todoWireData = Wire.data(id);
+    final todoVO = todoWireData.value as TodoVO;
+    final count = Wire.data(TodoDataParams.COUNT).value as int;
+
+    todoVO.completed = !todoVO.completed;
+
+    Wire.data(id, todoVO);
+    Wire.data(TodoDataParams.COUNT, count + (todoVO.completed ? -1 : 1));
+
     _save();
+
+    print('> TodoModel -> toggled: ' + todoVO.id + ' - ' + todoVO.text);
     return null;
   }
 
+  @override
+  void filter(TodoFilterValue filter) {
+    final todoList = Wire.data(TodoDataParams.LIST).value as List;
+    todoList.forEach((id) {
+      var todoWireData = Wire.data(id);
+      var todoVO = todoWireData.value as TodoVO;
+      var todoVisible = todoVO.visible;
+      switch (filter) {
+        case TodoFilterValue.ALL: todoVisible = true; break;
+        case TodoFilterValue.ACTIVE: todoVisible = !todoVO.completed; break;
+        case TodoFilterValue.COMPLETED: todoVisible = todoVO.completed; break;
+      }
+      if (todoVO.visible != todoVisible) {
+        todoVO.visible = todoVisible;
+        Wire.data(id, todoVO);
+      }
+    });
+    Wire.data(TodoDataParams.FILTER, filter);
+    print('> TodoModel -> filtered: ' + filter.toString());
+  }
+
+  @override
+  void clearCompleted() {
+    final todoList = Wire.data(TodoDataParams.LIST).value as List;
+    todoList.removeWhere((id) {
+      var todoWireData = Wire.data(id);
+      var todoVO = todoWireData.value as TodoVO;
+      if (todoVO.completed) {
+        todoWireData.remove();
+      }
+      return todoVO.completed;
+    });
+
+    _save();
+    print('> TodoModel -> clearCompleted: length = ' + todoList.length.toString());
+  }
+
   void _save() {
-    var list = <TodoVO>[];
+    var listToSave = <TodoVO>[];
     (Wire.data(TodoDataParams.LIST).value as List).forEach((id) =>
-      list.add(Wire.data(id).value)
+      listToSave.add(Wire.data(id).value)
     );
-    window.localStorage[LOCAL_STORAGE_KEY] = jsonEncode(list);
+    window.localStorage[LOCAL_STORAGE_KEY] = jsonEncode(listToSave);
   }
 }