+++ example/todo/web/model/dto/EditDTO.dart
@@ -1,0 +1,6 @@
+class EditDTO {
+  String text;
+  String id;
+
+  EditDTO(this.id, this.text);
+}